move validateHttpHeaders definition into cloud scheduler (#4022)
